requirements.txt
discord.py>=2.3.0
requests>=2.31.0
flask>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
schedule>=1.2.0
